🚸 Add buttons in the question bar
Add two buttons to the question bar for going back to home or showing
the answer. This makes it clear, that one can show the answer by
clicking the button.

While at it, move the question bar used by both the easy and hard
quizzes to a separate component to reduce duplicate code.